Refactoring. Clean multableState interface
